requirements/production.txt
Django>=3.0,<4.0 
psycopg2>=2.7,<3.0
gunicorn>=20.0,<21.0
pickle5>=0.0.9,<0.1.0
django-cors-headers>=3.2,<3.3
djangorestframework>=3.11,<4.0
django-fontawesome-5>=1.0,<2.0
antlr4-python3-runtime>=4.7,<4.8

